+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+### Fixed
+
+- Url in root route
+
 ## [1.3.1] - 2025-03-03
 
 ### Added